Erstellen von Player.java und der toInput Methode im Enum

+++ app/src/main/java/sirius/Actor.java
@@ -26,11 +26,11 @@ public abstract class Actor {
         return y;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
